+++ test/Poseidon/JannocoalesceSpec.hs
@@ -1,14 +1,17 @@
 {-# LANGUAGE OverloadedStrings #-}
 module Poseidon.JannocoalesceSpec (spec) where
 
-import           Poseidon.CLI.Jannocoalesce (makeNewJannoRows, mergeRow)
+import           Poseidon.CLI.Jannocoalesce (CoalesceJannoColumnSpec (..),
+                                             makeNewJannoRows, mergeRow)
 import           Poseidon.Janno             (CsvNamedRecord (..),
                                              JannoList (..), JannoRow (..),
                                              createMinimalSample, makeLatitude,
                                              makeLongitude)
 import           Poseidon.Utils             (testLog)
 
+import           Control.Monad.IO.Class     (liftIO)
 import qualified Data.HashMap.Strict        as HM
+import qualified Data.IORef                 as R
 import           SequenceFormats.Eigenstrat (EigenstratIndEntry (..), Sex (..))
 import           Test.Hspec
 
@@ -23,7 +26,10 @@ jannoTargetRow =
     in  row {
             jCountry = Just "Austria",
             jSite = Just "Vienna",
-            jDateNote = Just "dating didn't work"
+            jDateNote = Just "dating didn't work",
+            jAdditionalColumns = CsvNamedRecord $ HM.fromList [
+                ("AdditionalColumn2", "C")
+            ]
         }
 
 jannoSourceRow :: JannoRow
@@ -33,7 +39,11 @@ jannoSourceRow =
             jCountry   = Just "Austria",
             jSite      = Just "Salzburg",
             jLatitude  = makeLatitude 30.0,
-            jLongitude = makeLongitude 30.0
+            jLongitude = makeLongitude 30.0,
+            jAdditionalColumns = CsvNamedRecord $ HM.fromList [
+                ("AdditionalColumn1", "A"),
+                ("AdditionalColumn2", "B")
+            ]
         }
 
 jannoTargetRows :: [JannoRow]
@@ -57,49 +67,78 @@ testMergeSingleRow :: Spec
 testMergeSingleRow =
     describe "Poseidon.Jannocoalesce.mergeRow" $ do
         it "should correctly merge without fields and no override" $ do
-            merged <- testLog $ mergeRow jannoTargetRow jannoSourceRow [] False "Poseidon_ID" "Poseidon_ID"
+            cp <- liftIO $ R.newIORef 0
+            merged <- testLog $ mergeRow cp jannoTargetRow jannoSourceRow AllJannoColumns False "Poseidon_ID" "Poseidon_ID"
             jSite merged      `shouldBe` Just "Vienna"
             jGroupName merged `shouldBe` JannoList ["SamplePop"]
             jLatitude merged  `shouldBe` makeLatitude 30.0
             jLongitude merged `shouldBe` makeLongitude 30.0
+            jAdditionalColumns merged `shouldBe` CsvNamedRecord (HM.fromList [
+                    ("AdditionalColumn1", "A"),
+                    ("AdditionalColumn2", "C")
+                ])
         it "should correctly merge without fields and override" $ do
-            merged <- testLog $ mergeRow jannoTargetRow jannoSourceRow [] True "Poseidon_ID" "Poseidon_ID"
+            cp <- liftIO $ R.newIORef 0
+            merged <- testLog $ mergeRow cp jannoTargetRow jannoSourceRow AllJannoColumns True "Poseidon_ID" "Poseidon_ID"
             jSite merged      `shouldBe` Just "Salzburg"
             jGroupName merged `shouldBe` JannoList ["SamplePop2"]
             jLatitude merged  `shouldBe` makeLatitude 30.0
             jLongitude merged `shouldBe` makeLongitude 30.0
-        it "should correctly merge with fields and no override" $ do
-            merged <- testLog $ mergeRow jannoTargetRow jannoSourceRow ["Group_Name", "Latitude"] False "Poseidon_ID" "Poseidon_ID"
+            jAdditionalColumns merged `shouldBe` CsvNamedRecord (HM.fromList [
+                    ("AdditionalColumn1", "A"),
+                    ("AdditionalColumn2", "B")
+                ])
+        it "should correctly merge with fields selection and no override" $ do
+            cp <- liftIO $ R.newIORef 0
+            merged <- testLog $ mergeRow cp jannoTargetRow jannoSourceRow (IncludeJannoColumns ["Group_Name", "Latitude"]) False "Poseidon_ID" "Poseidon_ID"
             jSite merged      `shouldBe` Just "Vienna"
             jGroupName merged `shouldBe` JannoList ["SamplePop"]
             jLatitude merged  `shouldBe` makeLatitude 30.0
             jLongitude merged `shouldBe` Nothing
+            jAdditionalColumns merged `shouldBe` CsvNamedRecord (HM.fromList [
+                    ("AdditionalColumn2", "C")
+                ])
+        it "should correctly merge with negative field selection and no override" $ do
+            cp <- liftIO $ R.newIORef 0
+            merged <- testLog $ mergeRow cp jannoTargetRow jannoSourceRow (ExcludeJannoColumns ["Latitude"]) False "Poseidon_ID" "Poseidon_ID"
+            jSite merged      `shouldBe` Just "Vienna"
+            jGroupName merged `shouldBe` JannoList ["SamplePop"]
+            jLatitude merged  `shouldBe` Nothing
+            jLongitude merged `shouldBe` makeLongitude 30.0
+            jAdditionalColumns merged `shouldBe` CsvNamedRecord (HM.fromList [
+                    ("AdditionalColumn1", "A"),
+                    ("AdditionalColumn2", "C")
+                ])
         it "should correctly merge with fields and override" $ do
-            merged <- testLog $ mergeRow jannoTargetRow jannoSourceRow ["Group_Name", "Latitude"] True "Poseidon_ID" "Poseidon_ID"
+            cp <- liftIO $ R.newIORef 0
+            merged <- testLog $ mergeRow cp jannoTargetRow jannoSourceRow (IncludeJannoColumns ["Group_Name", "Latitude"]) True "Poseidon_ID" "Poseidon_ID"
             jSite merged      `shouldBe` Just "Vienna"
             jGroupName merged `shouldBe` JannoList ["SamplePop2"]
             jLatitude merged  `shouldBe` makeLatitude 30.0
             jLongitude merged `shouldBe` Nothing
+            jAdditionalColumns merged `shouldBe` CsvNamedRecord (HM.fromList [
+                    ("AdditionalColumn2", "C")
+                ])
 
 testCoalesceMultipleRows :: Spec
 testCoalesceMultipleRows = describe "Poseidon.Jannocoalesce.makeNewJannoRows" $ do
     it "should correctly copy with simple matching" $ do
-        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows [] False "Poseidon_ID" "Poseidon_ID" Nothing
+        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows AllJannoColumns False "Poseidon_ID" "Poseidon_ID" Nothing
         jCountry (newJ !! 0) `shouldBe` Just "Germany"
         jCountry (newJ !! 1) `shouldBe` Just "Austria"
         jCountry (newJ !! 2) `shouldBe` Nothing
     it "should correctly copy with simple matching and overwrite" $ do
-        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows [] True "Poseidon_ID" "Poseidon_ID" Nothing
+        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows AllJannoColumns True "Poseidon_ID" "Poseidon_ID" Nothing
         jCountry (newJ !! 0) `shouldBe` Just "Austria"
         jCountry (newJ !! 1) `shouldBe` Just "Austria"
     it "should throw with duplicate source keys" $ do
         let s = jannoSourceRows ++ [jannoSourceRows !! 1]
-        testLog (makeNewJannoRows s jannoTargetRows [] False "Poseidon_ID" "Poseidon_ID" Nothing) `shouldThrow` anyException
+        testLog (makeNewJannoRows s jannoTargetRows AllJannoColumns False "Poseidon_ID" "Poseidon_ID" Nothing) `shouldThrow` anyException
     it "should correctly copy with suffix strip" $ do
-        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows [] False "Poseidon_ID" "Poseidon_ID" (Just "_AB")
+        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows AllJannoColumns False "Poseidon_ID" "Poseidon_ID" (Just "_AB")
         jCountry (newJ !! 0) `shouldBe` Just "Germany"
         jCountry (newJ !! 1) `shouldBe` Just "Austria"
         jCountry (newJ !! 2) `shouldBe` Just "Austria"
     it "should correctly copy with alternative ID column" $ do
-        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows [] False "Poseidon_ID_alt" "Poseidon_ID" Nothing
+        newJ <- testLog $ makeNewJannoRows jannoSourceRows jannoTargetRows AllJannoColumns False "Poseidon_ID_alt" "Poseidon_ID" Nothing
         jCountry (newJ !! 2) `shouldBe` Just "Austria"